+++ jm-glass/src/components/index.tsx
@@ -1,4 +1,4 @@
 export { default as Chord } from './Chord';
 export { default as NoteSelector } from './NoteSelector';
 export { default as ChordSelector } from './ChordSelector';
-export { default as Tableture } from './Tableture';
+export { default as Tabulature } from './Tabulature';
